+++ src/jiratui/utils/adf.py
@@ -1,7 +1,7 @@
 import re
 
 from marklas import Transformer, parse_md, to_adf, to_md
-from marklas.ast import Inline, LinkMark, Mark, Paragraph, Text
+from marklas.ast import LinkMark, Mark, Text
 
 
 def convert_markdown_to_adf(content: str) -> dict:
@@ -30,12 +30,11 @@ def convert_adf_to_markdown(content: dict) -> str:
     return to_md(content)
 
 
-# the links extracted from a Markdown content
-links: list[dict] = []
-
-
 def extract_web_links_from_markdown(content: str) -> list[dict]:
     """Extracts Web links from a Markdown string.
+
+    This function will return a list without duplicated urls. If a URL appears more than once in the document then the
+    last occurrence is returned.
 
     Args:
         content: The Markdown string to extract Web Links from.
@@ -60,35 +59,32 @@ def extract_web_links_from_markdown(content: str) -> list[dict]:
 
     link_parser_transformer = Transformer()
 
-    @link_parser_transformer.register(Paragraph)
-    def _process_paragraph(node: Paragraph) -> Paragraph | None:
-        global links
+    @link_parser_transformer.register(Text)
+    def _process_text(node: Text) -> Text | None:
         raw_urls_regex = re.compile(
             r'https?://[^\s\[\]()]+|ftp://[^\s\[\]()]+|(?:www\.)[^\s\[\]()]+|(?<![\(\[])(?<![a-zA-Z]\()(?:(?<!\()\b(?:[a-z]{2,}\.)+(com|org|net|edu|gov|io|co|uk)\b)',
             re.MULTILINE,
         )
-        c: Inline
         m: Mark
-        for c in node.content:
-            if not isinstance(c, Text):
-                continue
-            link_text = c.text or ''
-            for m in c.marks:
-                if isinstance(m, LinkMark):
-                    links.append({'url': m.href, 'title': m.title or link_text})
-            # there may also be "raw" links in the text; i.e. w/o Markdown []() syntax. We need to extract these using a
-            # regex
-            if link_text:
-                for link in raw_urls_regex.finditer(link_text):
-                    if href := link.group():
-                        links.append({'url': href, 'title': href})
+        link_text = node.text or ''
+        for m in node.marks:
+            if isinstance(m, LinkMark):
+                links.append({'url': m.href, 'title': m.title or link_text})
+        # there may also be "raw" links in the text; i.e. w/o Markdown []() syntax. We need to extract these using a
+        # regex
+        if link_text:
+            for link in raw_urls_regex.finditer(link_text):
+                if href := link.group():
+                    links.append({'url': href, 'title': href})
+        # return the unmodified node
         return node
 
     if not content:
         return []
 
-    global links
-    links = []
+    links: list[dict] = []
     ast_doc = parse_md(content)
     link_parser_transformer(ast_doc)
-    return links[:] or []
+    # remove duplicated urls
+    links_by_urls = {link.get('url'): link for link in links}
+    return list(links_by_urls.values()) or []