+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Extract and display web links found in Markdown context. If links are found in the content then a collapsible section
 will appear in the Info tab's text panes (description, environment, etc.) so the use can open the desired
-links. By [@whyisdifficult](https://github.com/whyisdifficult) in https://github.com/whyisdifficult/jiratui/pull/346
+links. By [@whyisdifficult](https://github.com/whyisdifficult) in https://github.com/whyisdifficult/jiratui/pull/347
 
 ## [1.13.1] 2026-08-28